frontend fixes

+++ backend/src/main/java/ru/ibelan/test/backend/graphql/ContractQuery.java
@@ -34,7 +34,7 @@ public class ContractQuery implements GraphQLQueryResolver {
         return contractService.getRealPropertyTypes();
     }
 
-    public Float calcInsurancePremium(int insuranceAmount,
+    public Long calcInsurancePremium(int insuranceAmount,
                                       String realPropertyType,
                                       Integer year,
                                       Integer area,
